Move memory types into memory optimizer.
